녹음된 내용만 저장 후 makedraft에서 초안 생성(기능 구현x)

+++ src/main/java/syu/autobiography/spring/audioupload/repository/AudioRepository.java
@@ -3,5 +3,5 @@ package syu.autobiography.spring.audioupload.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import syu.autobiography.spring.entity.Drafts;
 
-public interface AudioRepository extends JpaRepository<Drafts, Integer> {
+public interface AudioRepository extends JpaRepository<Drafts, Long> {
 }
